Specify app colors in sRGB
I think I had sampled the colors in one space when measuring on-screen
colors for the ship, etc., but then was creating the UIColor in a
different space.  I had intended things to match but they didn't.

Colors for the game controller icons are wrong FWIW, but I'm not sure
if I'll bother to fix them.