Add "Lazarus" for gaining a extra life after losing the last ship
This is more likely now with the bonus for starting at a higher wave
and clearing it.  Still not overly likely, but the most common case is
likely to be dying to a UFO with one asteroid remaining and the UFO
quickly squeezing off a shot to destroy the last asteroid, or the
player's last shot clearing the wave just after they die.  Either way,
the last asteroid being cleared triggers the bonus and gives another
life.

Note that if the UFO doesn't fire and hit quickly, it won't work.
There's only about a four second window (in slow motion) after they
die before respawnOrGameOver happens.  The check for reserves happens
at the start of respawnOrGameOver, and gameOver will become true if
they have none.  Once gameOver is set, final asteroid removal will not
give the bonus.