Make sure global and local dates for counters are in sync
I don't really think this case should happen.  The dates for local and
global should be set together on a reset.  And if the person switches
to a different device after a reset, it should see the updated global
date with their local date behind.  I seem to have gotten into this
situation at least once though, maybe due to some sort of sync error.
Anyway, just ensure that dates are sync'd.