Add game controller color support
If there's a colored light on the controller, change it to match the
ship color.  Make player explosions flash a bit.  I also reworked the
color selection a bit, so they could replay the tutorial in a color of
their choice if desired.  Maybe add an achievement for doing it
not-in-blue?